+++ LIBRARY/src/bid64_to_uint64.c
@@ -108,7 +108,7 @@ bid64_to_uint64_rnint (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 20) {	// x >= 10^20 ~= 2^66.45... (cannot fit in 64 bits)
     // set invalid flag
@@ -338,7 +338,7 @@ bid64_to_uint64_xrnint (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 20) {	// x >= 10^20 ~= 2^66.45... (cannot fit in 64 bits)
     // set invalid flag
@@ -613,7 +613,7 @@ bid64_to_uint64_floor (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 20) {	// x >= 10^20 ~= 2^66.45... (cannot fit in 64 bits)
     // set invalid flag
@@ -793,7 +793,7 @@ bid64_to_uint64_xfloor (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 20) {	// x >= 10^20 ~= 2^66.45... (cannot fit in 64 bits)
     // set invalid flag
@@ -991,7 +991,7 @@ bid64_to_uint64_ceil (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 20) {	// x >= 10^20 ~= 2^66.45... (cannot fit in 64 bits)
     // set invalid flag
@@ -1204,7 +1204,7 @@ bid64_to_uint64_xceil (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 20) {	// x >= 10^20 ~= 2^66.45... (cannot fit in 64 bits)
     // set invalid flag
@@ -1422,7 +1422,7 @@ bid64_to_uint64_int (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 20) {	// x >= 10^20 ~= 2^66.45... (cannot fit in 64 bits)
     // set invalid flag
@@ -1611,7 +1611,7 @@ bid64_to_uint64_xint (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 20) {	// x >= 10^20 ~= 2^66.45... (cannot fit in 64 bits)
     // set invalid flag
@@ -1824,7 +1824,7 @@ bid64_to_uint64_rninta (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 20) {	// x >= 10^20 ~= 2^66.45... (cannot fit in 64 bits)
     // set invalid flag
@@ -2040,7 +2040,7 @@ bid64_to_uint64_xrninta (BID_UINT64 x
     if (C1 >= bid_nr_digits[x_nr_bits - 1].threshold_lo)
       q++;
   }
-  exp = x_exp - 398;	// unbiased exponent
+  exp = (int) (x_exp - 398);	// unbiased exponent
 
   if ((q + exp) > 20) {	// x >= 10^20 ~= 2^66.45... (cannot fit in 64 bits)
     // set invalid flag
